// src/components/molecules/dialog/components/dialog-container/components/dialog-resize/handlers/types/dialog-resize.handlers.return.type.ts
type DialogResizeHandlersReturnType = {
  handleBottomLeftCornerPointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
  handleBottomRightCornerPointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
  handleBottomSidePointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
  handleLeftSidePointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
  handleRightSidePointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
  handleTopLeftCornerPointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
  handleTopRightCornerPointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
  handleTopSidePointerDownEvent: (event: React.PointerEvent<HTMLDivElement>) => void;
};

export { DialogResizeHandlersReturnType };