// src/components/molecules/dialog/components/dialog-container/components/dialog-resize/handlers/types/top-left-corner-pointer-down-event-handler.props.type.ts
// Vendors
import { Dispatch, SetStateAction } from 'react';

type TopLeftCornerPointerDownEventHandlerPropsType = {
  event: React.PointerEvent<HTMLDivElement>;
  isFullscreen: boolean;
  minSize: { width: number; height: number };
  setSize: Dispatch<SetStateAction<{ width: number; height: number | null }>>;
  setTranslate: Dispatch<SetStateAction<{ x: number; y: number }>>;
  size: { width: number; height: number | null };
  translate: { x: number; y: number };
};

export { TopLeftCornerPointerDownEventHandlerPropsType };