// src/components/molecules/dialog/components/dialog-container/components/dialog-resize/handlers/types/top-side-pointer-down-event-handler.props.type.ts
// Vendors
import { Dispatch, SetStateAction } from 'react';

type TopSidePointerDownEventHandlerPropsType = {
  disabled?: boolean;
  event: React.PointerEvent<HTMLDivElement>;
  minSize: { width: number; height: number };
  setSize: Dispatch<SetStateAction<{ width: number; height: number | null }>>;
  setTranslate: Dispatch<SetStateAction<{ x: number; y: number }>>;
  size: { width: number; height: number | null };
  translate: { x: number; y: number };
};

export { TopSidePointerDownEventHandlerPropsType };